requirements.txt
Django==1.9.2
django-filter==0.13.0
psycopg2==2.6.1
